requirements.txt
flask
flask-cors
youtube-transcript-api
langchain
langchain-huggingface
langchain-community
langchain-text-splitters
faiss-cpu
python-dotenv
